requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
playwright
pandas
matplotlib
requests
beautifulsoup4
lxml
pdfplumber
